Improvements to interfaces and optimization of HydroUpdate_euler

See merge request dyablo/dyablo!367

+++ core/src/amr/LightOctree_storage.h
@@ -90,7 +90,7 @@ public:
 
   //! @copydoc LightOctree_base::getNdim()
   KOKKOS_INLINE_FUNCTION 
-  uint8_t getNdim() const
+  uint32_t getNdim() const
   {
       return ndim;
   }
